+++ codec_test.go
@@ -32,11 +32,11 @@ func TestMarshalUnmarshalPointer0(t *testing.T) {
 
 	var s = newSimpleStruct()
 	cdc := amino.NewCodec()
-	b, err := cdc.MarshalLengthPrefixed(s) // no indirection
+	b, err := cdc.MarshalSized(s) // no indirection
 	assert.NoError(t, err)
 
 	var s2 SimpleStruct
-	err = cdc.UnmarshalLengthPrefixed(b, &s2) // no indirection
+	err = cdc.UnmarshalSized(b, &s2) // no indirection
 	assert.NoError(t, err)
 	assert.Equal(t, s, s2)
 
@@ -46,11 +46,11 @@ func TestMarshalUnmarshalPointer1(t *testing.T) {
 
 	var s = newSimpleStruct()
 	cdc := amino.NewCodec()
-	b, err := cdc.MarshalLengthPrefixed(&s) // extra indirection
+	b, err := cdc.MarshalSized(&s) // extra indirection
 	assert.NoError(t, err)
 
 	var s2 SimpleStruct
-	err = cdc.UnmarshalLengthPrefixed(b, &s2) // no indirection
+	err = cdc.UnmarshalSized(b, &s2) // no indirection
 	assert.NoError(t, err)
 	assert.Equal(t, s, s2)
 
@@ -62,7 +62,7 @@ func TestMarshalUnmarshalPointer2(t *testing.T) {
 	var ptr = &s
 	cdc := amino.NewCodec()
 	assert.Panics(t, func() {
-		cdc.MarshalLengthPrefixed(&ptr) // double extra indirection panics.
+		cdc.MarshalSized(&ptr) // double extra indirection panics.
 		cdc.RegisterPackage(tests.Package)
 	})
 }
@@ -71,11 +71,11 @@ func TestMarshalUnmarshalPointer3(t *testing.T) {
 
 	var s = newSimpleStruct()
 	cdc := amino.NewCodec()
-	b, err := cdc.MarshalLengthPrefixed(s) // no indirection
+	b, err := cdc.MarshalSized(s) // no indirection
 	assert.NoError(t, err)
 
 	var s2 *SimpleStruct
-	err = cdc.UnmarshalLengthPrefixed(b, &s2) // extra indirection
+	err = cdc.UnmarshalSized(b, &s2) // extra indirection
 	assert.NoError(t, err)
 	assert.Equal(t, s, *s2)
 }