+++ tests/fuzz/binary/debug/main.go
@@ -13,6 +13,6 @@ func main() {
 	bz := []byte("\a\x1a\x05\x1a\x01\x80\xf7\x00")
 	cdc := amino.NewCodec()
 	cst := tests.ComplexSt{}
-	err := cdc.UnmarshalLengthPrefixed(bz, &cst)
+	err := cdc.UnmarshalSized(bz, &cst)
 	fmt.Printf("Expected a panic but did not. (err: %v)", err)
 }